Move Builder below MapTensor
